+++ prog/dftb+/lib_dftb/etemp.F90
@@ -24,21 +24,24 @@ module dftbp_etemp
   implicit none
   private
 
-  public :: Efilling, electronFill, Fermi, Gaussian, Methfessel
+  public :: Efilling, electronFill, fillingTypes
 
+  type :: TFillingTypesEnum
 
-  !> Definition of a type of broadening function - Fermi-Dirac in this case
-  integer, parameter :: Fermi = 0
+    !> Definition of a type of broadening function - Fermi-Dirac in this case
+    integer :: Fermi = 0
 
+    !> Definition of a type of broadening function - Gaussian in this case
+    integer :: Gaussian = 1
 
-  !> Definition of a type of broadening function - Gaussian in this case
-  integer, parameter :: Gaussian = 1
+    !> Definition of a type of broadening function - Methfessel-Paxton, for higher orders use
+    !> Methfessel + n as a value
+    integer :: Methfessel = 2
 
+  end type TFillingTypesEnum
 
-  !> Definition of a type of broadening function - Methfessel-Paxton, for higher orders use
-  !> Methfessel + n as a value
-  integer, parameter :: Methfessel = 1
-
+  !> Enumerated filling types.
+  type(TFillingTypesEnum), parameter :: fillingTypes = TFillingTypesEnum()
 
   !> Twice the machine precision
   real(dp), parameter :: epsilon2 = 2.0_dp * epsilon(1.0_dp)
@@ -106,7 +109,7 @@ contains
     @:ASSERT(size(kWeight) > 0)
     @:ASSERT(all(kWeight >= 0.0_dp))
 
-    @:ASSERT(distrib >= Fermi)
+    @:ASSERT(distrib >= fillingTypes%Fermi)
 
     ! If no electrons there, we are ready
     if (nElectrons < epsilon(1.0_dp)) then
@@ -174,7 +177,7 @@ contains
     nElec = electronCount(Ef, eigenvals, kT, distrib, kWeight)
     ! Polish resulting root with Newton-Raphson type steps
     if (abs(nElectrons - nElec) > elecTol) then
-      if (distrib == Fermi) then ! only derivs for Fermi so far
+      if (distrib == fillingTypes%Fermi) then ! only derivs for Fermi so far
         if (abs(derivElectronCount(Ef, eigenvals, kT, distrib, kWeight)) >= epsilon(1.0_dp)) then
           EfOld = Ef
           Ef = Ef - (electronCount(Ef, eigenvals, kT, distrib, kWeight) - nElectrons)&
@@ -238,8 +241,8 @@ contains
 
     w = 1.0_dp/kT
     electronCount=0.0_dp
-    if (distrib /= Fermi) then
-      MPorder = distrib - 1
+    if (distrib /= fillingTypes%Fermi) then
+      MPorder = distrib - fillingTypes%Methfessel - 1
       allocate(A(0:MPorder))
       allocate(hermites(0:2*MPorder))
       call Aweights(A,MPorder)
@@ -319,7 +322,7 @@ contains
 
     w = 1.0_dp/kT
     derivElectronCount=0.0_dp
-    if (distrib /= Fermi) then
+    if (distrib /= fillingTypes%Fermi) then
       call error("Fermi distribution only supported")
     else
       do ispin = 1, size(eigenvals,dim=3)
@@ -404,8 +407,8 @@ contains
     E0(:) = 0.0_dp
 
     ! The Gaussian and Methfessel-Paxton broadening functions first
-    if (distrib /= Fermi) then
-      MPorder = distrib - 1
+    if (distrib /= fillingTypes%Fermi) then
+      MPorder = distrib - fillingTypes%Methfessel -1
       allocate(A(0:MPorder))
       allocate(hermites(0 : 2 * MPorder))
       call Aweights(A, MPorder)